build: remove -fno-isolate-erroneous-paths-attribute
This flag is unsupported by clang and isolate-erroneous-paths-attribute is not enabled by default, thus it's may be the case in future when -O2 is used.
```Detect paths that trigger erroneous or undefined behavior due to a null value being used in a way forbidden by a returns_nonnull or nonnull attribute. Isolate those paths from the main control flow and turn the statement with erroneous or undefined behavior into a trap. This is not currently enabled, but may be enabled by -O2 in the future.```